c: implementation of skills
This commit contains the initial implementation of the
skills library in C. The C client depends on a single
library, hiredis, which is included as a submodule
in the third-party directory. The makefile for the
client supports the standard:
- make
- make clean
- make install (PREFIX=$INSTALL_PREFIX)

There is also a very simple test target and testing
utility in the `test` folder that is by no means
complete/exhaustive or working. It's basically
just a record of things that have been tested/working
and needs to be dramatically improved before
`make test` means anything. That said, `make test`
does compile and run the tests.

Documentation on the implementation of C skills
is sparse at the moment, but the headers/functions
themselves are implemented fairly accurately to the
spec and are well-documented.