atom: initial docker implementation
This is a major commit that does the following:

1. Renames "skills" naming scheme to "atom" based
naming scheme in C client

2. Implements docker containers for the atom build
system. The atom build system consists of two
main docker containers:
    (a) The "nucleus". This is based off of redis
Alpine and more or less runs redis. In the future
it will also have our CLI.
    (b) The "atom" container. This is auto-built
by circleCI and contains the latest atom libraries
as well as protobuf.

3. Sets up CircleCI to build the docker containers
and push them to dockerhub. Unit tests are also
run by spinning up both the nucleus and the atom
image using docker-compose.

See README for detailed instructions on use and
what's changed.