atom logging: Implement for C/C++
- atom_log APIs can now take a NULL element or NULL
context and still work. In the event of a NULL
element they'll just use a default element name
and for a NULL context they'll go ahead and make/cleanup
their own context for now.
- C/C++ APIs now use atom_log instead of fprintf/std::cout
- There's now a define in atom.h, ATOM_PRINT_LOGS, that
controls whether or not logs are also printed to stdout/stderr.
Both the C/C++ APIs respect this.
- C/C++ APIs now log the hostname under the "host" field in
the logs