.circleci: deploy more full set of tags
Was getting confusing in the build system not having consistency across
tags so that they end in -stock-amd64 and -stock-aarch64 vs latest/-aarch64.

Accordingly, deploy more tags using the target_tag_cmd field of the
CircleCI matrix s.t. we're deploying the usual tags such as
developmen-XYZ and development-XYZ-stock-amd64 s.t. when we run
automated builds we can easily use the matrix to link against
the correct one.