opengl, opengl+cuda: build graphics stage
Previously wasn't building graphics and this would have been
an issue.

Also, we were never testing from the actual stage being built
for the opengl-based images before; we were always testing
against the atom stage. Now, we test against a stage that's
a build arg and defaults to atom but can be set to graphics
for the opengl-based builds