base_build: use local registry
A bit of a complicated issue:
- buildx default driver uses a completely separate docker
container in order to run builds
- This docker container doesn't have access to locally cached
builds.
- The "default" docker driver which can access local builds
doesn't seem to cross-compile properly and is more machine-to-machine
dependent based on the version of Docker installed and is generally
not desirable to use.

As such, we need to have a way in which we can run a multi-stage
build that chains docker images without having to push them to
remote registries. To do this, we simply run a local registry
on localhost:5000 and use the new docker-compose-registry.yml
to do so. This docker-compose also links in the registry
folder s.t. it caches/saves images between boots which is nice.