Python element: add data_timeout_ms constructor arg
This argument will enforce that after we have established
a connection with the remote redis and we have sent it a command
we will wait for no longer than this number of milliseconds for data.

NOTE: This timeout is above and supersedes any timeouts set by
blocking redis calls, i.e. if you're using a N millisecond blocking
redis call you MUST instantiate your element with a data_timeout_ms
slightly greater than N. If data_timeout_ms is less than N the socket
will time out during the valid redis blocking window. This is by
design and desirable as we want to use this data_timeout_ms on the
linux/TCP socket level s.t. we can detect failed network connections.

The default timeout is 5s as there are not many blocking calls >5s
in known codebases. It's also generally recommended to implement
long-blocks for data with 1s blocks + polls and super long blocks
are generally not advised.