tests: Standardize element launch/cleanup, add guarantees
- Tests were previously subject to race conditions regarding
when an element would be up and ready to receive commands. This
was showing up on ARM builds with arbitrary infinite hangs. Now,
when launching elements, we standardize around calling
wait_for_healthy to prove the element is up and ready to receive
commands and thereby eliminating this race condition
- Standardization is now centered around using the non-blocking
command_loop call as opposed to manual process creaiton which is
significantly cleaner.
- Standardization is also joining + deleting on elements.