element: add in _clean_up function
The Python behavior/calling scope of __del__ is shaky and
can't be guaranteed as __del__ decrements a reference count
and cannot be forced to be run synchronously.

In the unit tests we want to be able to test the destructor
and ensure that streams are appropriately cleaned up. As such,
create an internal function, _clean_up() which is called by
__del__ but can also be called manually. Also, add an internal
variable to the Element s.t. this function isn't run twice if
we do manually call _clean_up() and then it's attempted to be
called again when __del__ is finally called by the garbage
collector.

The _clean_up function is meant for internal use and should really
never be called by users except in truly extraordinary
circumstances such as our test cases.