metrics: add, async + flush
- Adds capability to add data to a metric
- Adds capability to not execute the redis command when adding
data to a metric and instead leave the command buffered in a pipeline
in an async pool. This will be useful in the critical path where we
don't want the network overhead but still want to be using the metrics.
Data will be flushed to the redis server either the next time someone
calls add_metric with a synchronous execution call or when metrics_flush
is called
- Adds metrics_flush as above