metrics: make robust to potential dual-timestamp error
As documented in https://github.com/RedisTimeSeries/RedisTimeSeries/issues/492
if we have two timestamps that happen to be identical when trying
to add metrics we will run into an error from Redis. The behavior for this
is that the first write will go through and the second will be dropped.

Unit tests have been added to stress/check we're handling this properly
and this is a good enough workaround for now. We'll see where we get
with the issue -- not sure there is a fundamentally better solution
given the limitations of millisecond-level timing and uniqueness
required on time-series data