dashboards: working on programmatically generating dashboards
- Grafana configures its database + dashbaords using sqlite3, not
a good idea to try to save dashboards in the database itself.
- Will plan to use the dashboard HTTP API to generate system
high-level dashboards as well as per-element dashboards using
the data we have available through atom

This feature should also allow users to specify their own dashboard
configs that will get loaded in to the metrics system at boot if
done right.