RedisTimeSeries: working on fixing data-out-of-range bug
Redis time series pushed a commit that fixes the "Data outside of
time range" we're seeing in Grafana. This new commit however changed
the response of the TS.INFO command which then broke redistimeseries-py.

Upgrading redistimeseries-py then broke our semantics as they decided
to no longer subclass Redis but allow you to then pass your own connection.
This is a nice feature, but broke all of our code. Will need to refactor
the code for non-redis-timeseries commands to use client.redis instead
of the client directly.