element: Improve metrics_add_type API
This can replace the metrics_add_no_create function in
metrics_monitor.py. There's now a built-in API in atom to
create/update metrics in a single place as opposed to using
the create/add flow previously recommended which first creates
all known metrics at init and then adds them further on.

This comes at a performance hit as we will be re-creating the key
each time. The labels are only created/sent the first time which
is most of the performance hit so it's overall not terrible