metrics_monitor: make individual calls to redis to perform writes
- Previously we were pipelining all of the calls which was locking
up the redis for a good second when trying to write metrics. This
was suboptimal as it could cause a pretty large system hang. Overall
we should be OK with slightly worse performance in this monitoring
tool with the goal of minimizing overall impact on the system.
- Worst-case we can always put the monitoring metrics in a different
redis if need be but I don't think that's necessary yet