atom: turn off transactions on pipelines
It turns out we've been wrapping all of our pipeline calls in multi/exec
patterns. This causes the redis server to execute our _entire_ pipeline
before serving any other clients and was generally fine for most instances
in which we were using commands.

However, this fails on blocking-type list and set pop operations as these
will block the entire redis server until the blocking-type operation
completes. As such, when wrapping a BZPOP operation in our typical pipeline
we were immediately timing out (as specified per redis's docs on BLPOP)
instead of achieving the blocking behavior we desired.

As such, all redis pipelines now, by default, no longer wrap their data
in multi-exec type transaction blocks and simply exist to send all of the
data to the server in a single transaction. There's nothing in our redis
designs that would need the multi-exec so this is OK for now. If we wind
up needing multi-exec type guarantees in the future we can add in a
mechanism to get it back in some cases.