key deletion: Don't throw error on key missing
When deleting references it's typically OK to try to delete something
not there as we'd rather be over-cautious and risk calling it twice.
It's also potentially possible for a key to have expired, so the
deletion not happening is not a huge deal as long as at the end of the
day it's no longer taking up space in memory.

As such, modify the logic of _redis_key_delete() to no longer raise on
missing key, but rather just note that the call was unsuccessful and
return a list of the references that we did not successfully delete